Extracted primitive type related methods to separate utils class
